# Conflicts:
#	NenthomWeb/pom.xml
#	NenthomWeb/src/main/java/services/ConnectionUtil.java
#	NenthomWeb/src/main/java/servlets/Login_Servlet.java
#	NenthomWeb/src/main/java/servlets/Logout_Servlet.java
#	NenthomWeb/src/main/webapp/WEB-INF/web.xml
#	NenthomWeb/src/main/webapp/views/login.jsp
#	NenthomWeb/src/main/webapp/views/register.jsp
#	NenthomWeb/target/m2e-wtp/web-resources/META-INF/maven/spkt/NenthomWeb/pom.properties

+++ NenthomWeb/src/main/java/servlets/DeleteDiscount_Servlet.java
@@ -8,13 +8,17 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
+import utils.CSRFUtil;
 
 @WebServlet("/servlets/DeleteDiscount_Servlet")
 public class DeleteDiscount_Servlet extends HttpServlet {
     private DiscountDAO discountDAO = new DiscountDAO();
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    	if (!CSRFUtil.isValid(request)) {
+            request.getRequestDispatcher("/views/csrf_error.jsp").forward(request, response);
+            return;
+        }
         int discountID = Integer.parseInt(request.getParameter("discountID"));
 
         if (discountDAO.deleteDiscount(discountID)) {
